# Conflicts:
#	sumaris-core/src/main/java/net/sumaris/core/dao/referential/IEntityWithJoinSpecifications.java
#	sumaris-core/src/main/java/net/sumaris/core/dao/referential/IEntityWithStatusSpecifications.java
#	sumaris-core/src/main/java/net/sumaris/core/model/social/EventLevelEnum.java
#	sumaris-core/src/main/java/net/sumaris/core/model/technical/history/ProcessingHistory.java
#	sumaris-core/src/main/java/net/sumaris/core/service/social/UserEventService.java
#	sumaris-core/src/main/java/net/sumaris/core/service/social/UserEventServiceImpl.java
#	sumaris-core/src/main/java/net/sumaris/core/vo/social/UserEventFilterVO.java
#	sumaris-core/src/main/java/net/sumaris/core/vo/social/UserEventVO.java

+++ sumaris-core/src/main/java/net/sumaris/core/vo/data/batch/TempDenormalizedBatchVO.java
@@ -26,12 +26,14 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.FieldNameConstants;
 
+import java.math.BigDecimal;
+
 @Data
 @FieldNameConstants
 @EqualsAndHashCode
 public class TempDenormalizedBatchVO extends DenormalizedBatchVO {
 
-    private Double elevateFactor;
+    private BigDecimal elevateFactor;
 
     //private Double contextWeight;
     //private Double sumChildContextWeight;
